Fix missing imports to Viewis, pass as params from controller invocations
